Move RunConfig{} to a separate package
